requirements.txt
Django
markdown
djangorestframework
django-cors-headers